+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.1-SNAPSHOT
 - walk inheritance chain to find implementation
-- fix nil inheriting Object implementations
+- fix: nil should not inherit Object implementations
+- fix: walk superclass chain even with :extend-via-metadata
 
 # 1.0.0 - 10th September 2021
 - initial public release